Added the functionality to the checkboxes
When a checkbox is clicked now it marks the task as completed and colours it green. If a task is expired it colours the checkbox in red.

+++ taskify/tasksApp/models.py
@@ -126,7 +126,8 @@ class Task(models.Model):
         return self.title
 
     def delete(self, *args, **kwargs):
-        self.reminder.delete()
+        if self.reminder:
+            self.reminder.delete()
         super().delete(*args, **kwargs)
 
     class Meta:
